+++ jest.config.json
@@ -1,0 +1,5 @@
+{
+    "preset": "ts-jest",
+    "testEnvironment": "node",
+    "rootDir": "./src"
+}
