+++ package.json
@@ -7,7 +7,7 @@
     "start:dev": "nodemon",
     "start:prod": "tsc && node ./dist/index.js",
     "lint": "eslint . --ext .ts",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest"
   },
   "repository": {
     "type": "git",
@@ -20,12 +20,17 @@
   },
   "homepage": "https://github.com/dumbus/crud-api#readme",
   "devDependencies": {
+    "@types/jest": "^29.2.5",
     "@types/node": "^18.11.18",
+    "@types/supertest": "^2.0.12",
     "@types/uuid": "^9.0.0",
     "@typescript-eslint/eslint-plugin": "^5.48.1",
     "@typescript-eslint/parser": "^5.48.1",
     "eslint": "^8.31.0",
+    "jest": "^29.3.1",
     "nodemon": "^2.0.20",
+    "supertest": "^6.3.3",
+    "ts-jest": "^29.0.5",
     "ts-node": "^10.9.1",
     "typescript": "^4.9.4"
   },
