--- src/app/database/database.json
@@ -1,1 +1,0 @@
-[]
