+++ src/app/server.ts
@@ -5,7 +5,7 @@ import { requestListener } from './requestListener';
 dotenv.config();
 
 const port = process.env.PORT || 3000;
-const server = createServer(requestListener);
+export const server = createServer(requestListener);
 
 const startServer = () => {
     server.listen(port, () => console.log(`Server started on port ${port}`));
