+++ src/app/validators/validateUserId.ts
@@ -2,7 +2,7 @@ import { ServerResponse } from 'node:http';
 import { validate as validateUuid } from 'uuid';
 
 import { StatusCodes, ErrorMessages } from '../utils/messages';
-import * as database from '../database/database.json';
+import { database } from '../controllers/databaseController';
 
 const validateUserId = (userId: string, res: ServerResponse) => {
     try {
