+++ tsconfig.json
@@ -3,6 +3,7 @@
         "module": "commonjs",
         "target": "es6",
         "resolveJsonModule": true,
+        "esModuleInterop": true,
         "outDir": "dist"
     },
     "include": ["src"],
